+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## v0.5.24:
+
 ## v0.5.22:
 * Change  : Replace the use of cutlery's notifying_action with Chef 11's use_inline_resources method.
 * Bug     : Ensure that changes to custom resources are updated on the glassfish server if required.